requirements.txt
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.4
PyYAML==6.0.2
scipy==1.14.1
tifffile==2025.9.20
